# AstraServerGo.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AstraServerGo",
    "description": "基于 `AstraScheduleServerGo` 源码生成的 OpenAPI 3.0 文档。\n\n- 共 55 个 API 端点，覆盖客户端课表、天气、WebSocket、管理后台认证、用户管理、学校/年级/班级管理、配置管理、自动任务、倒数日、备份导入导出等功能。\n- JWT 认证接口使用 `jwtAuth` 安全方案（通过 POST /web/auth/login 获取 Token）。\n- 写入接口同时支持 BasicAuth（兼容旧版客户端）和 JWT + 密码验证。\n- 部分接口返回结构是动态 JSON（`map[string]interface{}`），已尽量按实际行为建模。\n",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/web/structure": {
      "get": {
        "summary": "获取学校-年级-班级结构",
        "operationId": "getStructure",
        "tags": ["web-menu"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "结构树",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/StructureNode" }
                }
              }
            }
          },
          "500": {
            "description": "错误响应",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorWithErrorField" }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/web/schools": {
      "post": {
        "summary": "创建学校",
        "operationId": "createSchool",
        "tags": ["web-structure"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200Message" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "未认证" },
          "409": {
            "description": "学校已存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/schools/{school}": {
      "delete": {
        "summary": "删除学校",
        "operationId": "deleteSchool",
        "tags": ["web-structure"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200Message" }
              }
            }
          },
          "401": { "description": "未认证" }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/schools/{school}/grades": {
      "post": {
        "summary": "创建年级",
        "description": "创建年级并自动生成默认科目和作息表。",
        "operationId": "createGrade",
        "tags": ["web-structure"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200Message" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "未认证" },
          "409": {
            "description": "年级已存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/schools/{school}/grades/{grade}": {
      "delete": {
        "summary": "删除年级",
        "operationId": "deleteGrade",
        "tags": ["web-structure"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200Message" }
              }
            }
          },
          "401": { "description": "未认证" }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/schools/{school}/grades/{grade}/classes": {
      "post": {
        "summary": "创建班级",
        "description": "创建班级并自动生成默认课表和客户端配置。",
        "operationId": "createClass",
        "tags": ["web-structure"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200Message" }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          },
          "401": { "description": "未认证" },
          "409": {
            "description": "班级已存在",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DetailResponse" }
              }
            }
          }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
    "/web/schools/{school}/grades/{grade}/classes/{class_number}": {
      "delete": {
        "summary": "删除班级",
        "operationId": "deleteClass",
        "tags": ["web-structure"],
        "parameters": [
          {
            "name": "school", "in": "path", "description": "学校标识", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "grade", "in": "path", "description": "年级", "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "class_number", "in": "path", "description": "班级", "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status200Message" }
              }
            }
          },
          "401": { "description": "未认证" }
        },
        "security": [{ "jwtAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StructureNode": {
        "type": "object",
        "properties": {
          "text": { "type": "string" },
          "children": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StructureNode" },
            "nullable": true
          }
        }
      },
...
      "ErrorWithErrorField": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "DetailResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        }
      },
...
      "Status200Message": {
        "type": "object",
        "required": ["status", "message"],
        "properties": {
          "status": { "type": "integer", "example": 200 },
          "message": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "（已废弃，仅供旧版客户端兼容）用户名：`AstraSchedule` 或 `ElectronClassSchedule`，密码：配置文件 `secret.token`"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，通过 POST /web/auth/login 获取。写入接口还需在 X-Verify-Password 头中提供密码。"
      },
      "internalAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Secret",
        "description": "内部 API 密钥，用于服务间调用。配置文件 [internal].secret"
      },
      "regToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Reg-Token",
        "description": "注册 JWT 令牌，由 reg-to 签发，有效期 10 分钟。包含完整注册信息。"
      }
    }
  },
  "servers": [
    {
      "url": "https://{hostvar1}.getastra.cn"
    }
  ],
  "security": []
}